requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
autobahn==21.2.1
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.8
constantly==15.1.0
cryptography==36.0.0
daphne==3.0.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.9
django-crispy-forms==1.13.0
django-filter==21.1
django-import-export==2.7.0
docker==5.0.3
et-xmlfile==1.1.0
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
MarkupPy==1.14
msgpack==1.0.3
odfpy==1.4.1
openpyxl==3.0.9
psycopg2==2.7.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==21.0.0
python-decouple==3.5
pytz==2021.3
PyYAML==6.0
requests==2.26.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
Twisted==21.7.0
txaio==21.2.1
typing_extensions==4.0.0
urllib3==1.26.7
websocket-client==1.2.1
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
zope.interface==5.4.0
